# microservices/analytics/audio-analytics/src/audio-analytics-api/data/api/openapi.yaml
...
openapi: "3.0.3"
info:
  title: "Audio Analytics API"
  version: "1.0.0"
  description: |
    REST API providing AI-powered audio processing: speech recognition (ASR),
    text translation (T2T), and speech synthesis (TTS).

    ## Services

    | Service | Endpoint prefix | Description |
    |---------|----------------|-------------|
    | ASR | `/transcriptions/` | Whisper-based speech-to-text (file or live WebSocket streaming) |
    | T2T | `/translations/` | Opus/NLLB text translation (batch supported) |
    | TTS | `/tts/` | MeloTTS speech synthesis; returns raw PCM audio |

    ## Audio formats (ASR input)
    - WAV, 16-bit PCM; 16 kHz mono recommended; max 25 MB

    ## TTS output
    - Supported sample rates: 44100, 22050, 16000 Hz (auto-resampled)

    ## WebSocket streaming sequence

    ```
    App                                           Server
     |===POST /transcriptions/create================>|
     |<==200 {session_id, state: asr_initialized}====|
     |
     |===WS /stream=================================>|
     |<==  {state: connection_established}===========|
     |
     |===WS {transcriptions_session_audio} (loop)===>|
     |<===WS {transcript.event,     speech_start}====|
     |<===WS {transcript.text.delta, ...}============|  (zero or more)
     |<===WS {transcript.event,     speech_end}======|
     |<===WS {transcript.text.done,  ...}============|
     |
...
    ```

    ## Authentication
    None required.



...
servers:
...
security: []
paths:
...
  /transcriptions/flush:
    post:
      tags: ["Transcriptions"]
      summary: "Flush buffered audio and force immediate transcription"
      description: |
        Forces the ASR engine to process any audio currently buffered in the
        VAD/silence-detection pipeline and emit a transcription result
        immediately, without waiting for the next speech-end event.

        Use this endpoint when you want to retrieve a partial result mid-stream
        (e.g. the user pressed a "done speaking" button) without fully closing
        the session.

        The session remains open after a flush and can continue receiving audio.
      operationId: "flushTranscription"
      x-swagger-router-controller: TranscriptionsFlush
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                session_id:
                  type: string
                  description: "Session ID to flush. If omitted, the only active session is used."
                  example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        "200":
          description: "Flush triggered successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "flushed"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /translations/translate:
    post:
      tags: ["Translations"]
      summary: "Translate text between languages"
      description: |
        Translates text from a source language to a target language.
        
        ## Features
        - Batch translation (multiple texts in one request)
        - Automatic language detection (if source not specified)
        - Custom parameters for fine-tuning
        
        ## Supported Languages
        Query `/translations/models` to see available language pairs.
        
        ## Example Request
        ```json
        {
          "text": ["Hello world", "How are you?"],
          "model": "opus-en-es",
          "source_language": "en",
          "target_language": "es"
        }
        ```
        
        ## Example Response
        ```json
        {
          "translations": [
            {
              "translated_text": "Hola mundo",
              "target_language": "es",
              "source_language": "en"
            },
            {
              "translated_text": "¿Cómo estás?",
              "target_language": "es",
              "source_language": "en"
            }
          ]
        }
        ```
      operationId: "translateText"
      x-swagger-router-controller: TranslationsTranslate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TranslationRequest"
            example:
              text: ["Hello world", "How are you?"]
              model: "opus-en-es"
              source_language: "en"
              target_language: "es"
      responses:
        "200":
          description: "Translation successful"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationResponse"
        "400":
          description: "Bad request - invalid input"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /tts/devices:
    get:
      tags: ["TTS"]
      summary: "List available audio output (playback) devices"
      description: "Returns available audio output devices (speakers) on the host system."
      operationId: "listTTSDevices"
      x-swagger-router-controller: TTSDevices
      responses:
        "200":
          description: "List of available audio output devices"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AudioOutputDevice"
        "500":
          description: "Internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      description: "Standard error response structure"
      properties:
        error:
          type: object
          properties:
            message:
              type: string
              description: "Human-readable error message"
              example: "Invalid request, missing 'file' parameter"
            type:
              type: string
              description: "Error type or category"
              example: "invalid_request_error"
            param:
              type: string
              nullable: true
              description: "Parameter related to the error (if applicable)"
              example: "file"
            code:
              type: string
              nullable: true
              description: "Error code (if applicable)"
              example: null
          required:
            - message
            - type
      required:
        - error
...
    TranslationRequest:
      type: object
      description: "Translation request"
      properties:
        text:
          type: array
          items:
            type: string
          description: "Array of text strings to translate"
          example:
            - "Hello world"
            - "How are you?"
        model:
          type: string
          description: "Translation model ID"
          example: "opus-en-es"
        source_language:
          type: string
          description: "Source language code (ISO 639-1). If omitted, may be auto-detected."
          example: "en"
        target_language:
          type: string
          description: "Target language code (ISO 639-1)"
          example: "es"
        parameters:
          description: "Optional parameters for translation"
          oneOf:
            - type: object
              additionalProperties: {}
            - type: array
              items:
                $ref: "#/components/schemas/KeyValueParam"
          example:
            - key: "formality"
              value: "informal"
      required:
        - text
        - model
        - target_language
    TranslationResponse:
      type: object
      description: "Translation result"
      properties:
        translations:
          type: array
          items:
            type: object
            properties:
              translated_text:
                type: string
                description: "Translated text"
                example: "Hola mundo"
              target_language:
                type: string
                description: "Target language code"
                example: "es"
              source_language:
                type: string
                description: "Source language code (detected or specified)"
                example: "en"
            required:
              - translated_text
          description: "Array of translation results (same order as input)"
      required:
        - translations
...
    KeyValueParam:
      type: object
      description: "Key-value pair for additional parameters"
      properties:
        key:
          type: string
          description: "Parameter name"
          example: "sampling_rate"
        value:
          type: string
          description: "Parameter value"
          example: "16000"
      required:
        - key
        - value
    AudioOutputDevice:
      type: object
      description: "Audio output (playback) device information"
      properties:
        index:
          type: integer
          description: "Device index"
          example: 0
        name:
          type: string
          description: "Device name"
          example: "LEMANS-EVK: - (hw:0,0)"
        default_samplerate:
          type: number
          description: "Default sample rate in Hz"
          example: 48000
        max_output_channels:
          type: integer
          description: "Maximum number of output channels"
          example: 2
      required:
        - index
        - name
...